feat: 로또번호 매칭 후 등수 매기는 기능

+++ src/main/java/lotto/model/lotto/generator/LottoGenerator.java
@@ -4,6 +4,6 @@ import lotto.model.lotto.Lottos;
 
 public interface LottoGenerator {
 
-    Lottos generate(long purchaseAmount);
+    Lottos generate(int purchaseAmount);
 
 }
